+++ CHANGELOG.md
@@ -4,6 +4,14 @@ Please note that this is not a complete list of every change. It will be continu
 
 Changes
 -----------------------
+**--04/17/2026--**
+
+* server:<br />drop now redundant `sv_cl_fps` cvar and don't force clients to accept downloads
+
+**--04/16/2026--**
+
+* cmake:<br />fix mingw
+
 **--04/15/2026--**
 
 * qcommon:<br />share x86 specific optimization with others